add invalid auth tests
